Various improvements.
Adding the functionalities to share the actual background image, to enable or disable the download of a new background image at each start.
Creation of "SettingsManager" class to store and retrive the application setting from the UserDefault.
Modification of the "SuperViewController" class, the class from each controller in this application should inherit from. The modification allows to download the new background without the blur effect, which is later applied using a UIVisualEffectView with UIBlurEffect.
Replacement of the deprecated "keyWindow", to always shows the loadingView on top, with a new version not deprecated.